#116 Change local and CI documentation build to use Sphinx
Sadly it seems very hard to locally mix a docs/requirements.txt with poetry when building both locally and on ReadTheDocs. So for the time being, they use different build methods.

But unlike before, the Sphinx version in poetry.toml and docs/requirements.txt are the same. This should allow to detect possible documentation build error before they make it to ReadTheDocs.