#113 Remove temporary directory in the output of a git analysis
When analyzing a remote git repository a temporary folder is created. The output contained the name of this temporary folder. This commit removes the temporary directory from the displayed output path.
Therefore, a new dataclass called PathData is introduced to contain the source path and temporary directory information of a path.